Fixed exception displayed on console after loading any valid json file
In HomePage.vue, fileUploaded variable was being set to True in the the
uploadStatus() method which triggers the AnnotationsPage component and
which expects json document in the json variable. Since the actual json
document gets loaded only after readFile() method, code throws an
exception in the background. While it does not impact the functionality
of application, it's still an irritating exception to have in the
console. With this fix, i am now marking this flag as True within the
readFile() method.